added datatype folder
added Volunteer data type
added Equipment data type
(sorry for the get and set architecture iain but i find it is so
standard and basic that it doesn’t make sense not to do it, even though
it is kind of archaic)

would you guys like me to implement cloneable so we can  implement the
volunteer and equipment classes a generic objects?

also I’m unsure if isRental and rentalCost adds to the effectiveness of
the current implementation.

note: it is worth noting that i am pretty rusty at this whole java
thing :p i don’t think i made any errors but if i have let me know.