stuff! things!
-can now access the add volunteer menu from the main menu
-can now add availability (so that the user can see it in the drop down
box (just try it out you will see))
-have yet to implement the adding in a gregorian way but have set it up
in a way that we will be able to as soon as we decide HOW we want to do
it, this is more of a design issue than a coding issue! (see comments
in code)
-on a submit it is now prepared to add a volunteer will the names,
email, and phone data initialized!
-(unsure what sort of list it will be added just that it is all hopped
up and ready to go ;) )
-once again though waiting on how we implement availability so we can
add that to the constructor.
-fb or text me if you have any questions, GAHD i hope this sinks
without having to merj.